Align plot to ggheatmap

Align a ggheatmap object to another ggplot that shares either its columns or rows to form complex panels. This function is called internally by add_tracks, and also exported by ggheatmap.

align_to_hm(
  gghm,
  gplot,
  pos = "bottom",
  newplt_size_prop = 0.5,
  legend_action = NULL,
  tag_level = "new"
)

Arguments

gghm

A object with class ggheatmap, generated by ggheatmap().

gplot

Another object of class ggplot, that shares either columns (if pos='top' or pos='bottom') or rows (if pos='left') with the heatmap. Note: to create the gplot, make sure you use gghmData() to get the data, or alternatively get_rowLevels() or get_colLevels().

pos

One of 'top', 'bottom' or 'left'.

newplt_size_prop

A float between 0 and 1, that indicates the proportion of the height (if pos='top' or pos='bottom') or width (if pos='left') that the new plot will occupy in the panel.

legend_action

A string specifying how guides should be treated in the layout. See: guides in patchwork::plot_layout().

tag_level

A string ('keep' or 'new') to indicate how auto-tagging should behave. See patchwork::plot_annotation.

Value

A joined panel object of class ggheatmap.